Refactoring of the NewHealthServerWithListener
Sometimes the test requires a gRPC server with only
a health service registered. It doesn't matter if
there are some kind of interceptors on it or not.
It is only required to check the flow is working
and the request reaches the service.
Now it is not required to explicitly stop the server
as it would be terminated automatically at the end
of the test execution. And as there are no other
use of the returned server the function doesn't return
it anymore.