ci: Fix binaries job being broken by missing Makefile target
With commit 5bf73d7a0 (Makefile: Drop assembly targets, 2021-04-28),
we've dropped both the assemble targets and the binaries target which
uses them. While the commit (probably validly) claims that no downstream
distributions or documentation use those targets, it was overlooked that
Gitaly itself does use it: when creating tags, the "binaries" CI job
compiles Gitaly with `make binaries` and then makes resulting binaries
available as artifacts, which is now broken.

Instead of bringing back the assemble targets, this commit instead
implements the job on top of `make build`: it does the same than `make
binaries` did, with the minor difference that the directory where
binaries end up in changes and that we don't create checksums.

Checksumming is thus now moved into CI and paths are adjusted to
accomodate for those changes.