Move CreateTag to the gittest package
CreateTag function is used to create tags and should
be moved into the gittest package. It not only more
consistent, but also breaks a circular dependency that
occurs if we replace MustRunCommand with the gittest.Exec
function. This replacement is done as well to break
dependency on the global config.Config variable.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699