git: Provide new RepositoryExecutor interface
In many cases, we currently pass around the `localrepo.Repo` structure
directly. But due to a cyclic import dependency between the localrepo
and the catfile packages, it's not possible to implement the catfile in
terms of the localrepo package.

To solve this, provide a new RepositoryExecutor interface in the git
package, which provides the ability to execute Git commands in a
specific repository.