gittest: Implement helper to write trees
There is no central helper function which allows us to write Git tree
objects into a repository. This commit thus implements a new function
`gittest.WriteTree()`, which, given a list of tree entries, will format
the tree object and use git-mktree(1) to write it into the object
database.

This commit also adds test infrastructure for the gittest package and a
single testcase for `WriteTree()`. Because of the design of our test
helpers which always cause the testcase to fail in case anything goes
wrong, it's not really possible to exercise failure code paths.