gittest: Refactor interface to write commits
The current interface to write commits is leaving a lot to be desired.
It's inflexible, we have multiple functions which all do some specific
things which others don't, and in general it's not a breeze to use.

As a first step to improve it, this commit refactors `CreateCommit()`:
instead of an options structure it now receives callback-like options
and it is renamed to `WriteCommit()` to bring it in line with the other
object-writing test helpers.