Include backchannel ID with transactions
Gitaly currently passes the backchannel ID with the PraefectServer
info. With the backchannel voting enabled by default, the server info
will no longer be needed. Passing the backchannel separately along
the transaction requires us to plumb the value through which requires
more boileplate for no real benefit.

This commit passes the backchannel ID with the transaction information
through the hooks. With the information passed in the transaction,
this allows us to remove the PraefectServer info entirely in a later
release and start using the backchannel id from the transaction directly.

Right now, the backchannel ID is included in the same Transaction struct
used by Praefect but the field is actually filled in by the Gitaly. It makes
sense to separate this later to better convery which service fills in which
information. Right now though, this should suffice for passing the information
through. Once this is in place, we can later refactor the code as we want without
affecting the API between Praefect, Gitaly and the hooks.