objectpool: Improve error message when fetching into pool fails
When git-fetch(1) fails to update the object pool in
`FetchFromOrigin()`, then the resulting error is rather unhelpful. It's
got no context, no error code and would typically only say that the
command failed with an exit code.

Improve the error adding stderr, using a proper error code and providing
context.