Replace MustRunCommand with gittest.Exec in upload_pack_test.go
To break dependency on the global config.Config variable the
git commands should be executed with gittest.Exec.
The functions require a config.Cfg to be passed as incoming parameter to
use proper git binary for the command execution. This change modifies
all related functions to pass through config.Cfg to the place of
the function replacement.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699