Replace MustRunCommand with gittest.Exec|ExecStream in service directory
To break dependency on the global config.Config variable the
git commands should be executed with gittest.Exec or gittest.ExecStream.
The functions require a config.Cfg to be passed as incoming parameter to
use proper git binary for the command execution.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699