gittest: Replace `WriteCommitWithName()` with `WriteCommit()`
The new `WriteCommit()` interface allows to implement most usecases for
which we previously had specialized helper functions. One usecase that
cannot yet be implemented is `WriteCommitWithName()` because it is still
lacking the ability to use a custom tree as root tree.

Implement a new `WithTreeEntries()` option and convert existing callers
of `WriteCommitWithName()` to use it. Drop the now unused function.