catfile: Inject Git command factory into cache
Right now, catfile commands get spawned via a global cache variable
where the `New()` function makes use of the cache. This is an inversion
of dependencies given that the cache should know to populate itself, not
the other way round. This dependency inversion also necessitates that
there is only a single cache, which is represented by the fact that it's
a global variable in the catfile package.

Inject the command factory into the cache as a preparatory step to fix
this technical debt. This will allow the cache to construct Git commands
itself instead of relying on the command factory passed in `New()`.