catfile: Make BatchCache type public
While we already have a Cache interface, we will eventually move
Prometheus metrics into the cache instead of relying on global
variables. To do this, we need to expose the `BatchCache` type and
return it from the (not yet public) `newCache()` function such that
users of it can register metrics.

Make the structure public, but hide all functions except for `Evict()`
to clean the cache and `BatchProcess()` to spawn a new one.