service: Set up catfile cache as global dependency
We have all prerequisites in place to start injecting the catfile cache
as dependency into its users, which will eventually let us get rid of
the global cache variable.

Implement a new `NewCache()` function which takes as input the Gitaly
configuration and command factory and inject it into service
dependencies.