localrepo: Provide test helper to construct repos
It's really hard right now to add additional dependencies on the
localrepo constructor given that it's used all over the place. To
address one part of the problem, introduce a new test helper which
constructs a `localrepo.Repo` given only a Gitaly config and a
repository interface -- all other dependencies get created ad-hoc.

Ideally, this would not be hosted in the localrepo package itself but in
the gittest package, but this doesn't work due to cyclic dependencies.
But given both the function name `NewTestRepo()` and that the function
requires a `testing.T` interface as first parameter, it shouldn't really
be possible to use it in production code.