services: Provide helper function to construct localrepos
With the same reasoning as layed out in the preceding commit, introduce
a set of helper functions for service servers which construct a
localrepo given only a repository interface -- all other dependencies
are taken from the server. This change will eventually allow us to
easily add additional dependencies to the localrepo implementation
without having to change hundreds of `localrepo.New()` calls.