repository: Fix secondaries being out-of-date after `Create()` RPC
All the `Create()` RPC does is to create leading directories for the
repository that is about to be created and then to call git-init(1).
While git-init(1) does run with hooks enabled, the reference-transaction
hook never executes given that no reference gets written. As a result,
no transactional voting will happen in the RPC and Praefect would
consider the transaction as potentially failed and replicate the change
to secondaries if Praefect was to set up transactions for the call.

Given that Praefect never does set up transactions for calls which
create new repositories because it treats them specially, this bug isn't
ever hit right now. But this will change soon, and to prepare for that
change we now start to cast manual votes when `Create()` has initialized
the repository.

Changelog: fixed