Create db records for secondaries that completed CreateRepo transaction
CreateRepository currently assumes that none of the secondaries participate
in a transaction when CreateRepository is called. As such, it doesn't create
`storage_repositories` records for them, indicating they didn't participate
in the transaction and are still outdated while waiting for a replication job
to be applied.

The secondaries not participating in a repository creating transaction causes
a lot of load on the replication queue when seeding a Gitaly Cluster with a lot
of data.

This is about the change, and the secondaries are going to participate in the
transaction when creating a repository. To allow for that, this commit changes
CreateRepository to accept `updatedSecondaries` and `outdatedSecondaries` to
differentiate which secondaries have the repository on the disk after finishing
the transaction. The updated secondaries have a `storage_repositories` record
created for them, and the complete set of secondaries have assignment records
created for them if variable replication factor is in use.