Simplify transaction ID generation by using montonically increasing IDs
Praefect previously had to use random transaction IDs to avoid collisions
when a Praefect restarts with transactions still in-flight in Gitalys. With
transactions now flowing through the backchannels, we can start using
monotonically increasing IDs and simplify the generation code. Gitaly only
sends transactions back via the connection they came from, so any in-flight
transactions would still go to the correct Praefect even after the socket
handover following a graceful upgrade. A Praefect process itself uses monotonic
IDs, so collisions are no longer possible. With the random IDs, there's a
possibility of generating an ID that collides with another transaction ID in the
same Praefect process. Chance of this happening is extremely low but monotonic
IDs are simpler and don't have this problem, so we might as well switch to them.