wiki: Remove DeletePage RPC
The Wiki service has functional duplication with the OperationService
RPCs. This means that if the clients are refactored, the wiki RPCs can
be removed. This change removes the DeletePage RPC as the client that
was using it, is no more: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/57106.

The proto is removed too, as the client is unable to call the RPC and
doesn't use it anymore.