ssh: Fix secondaries being out-of-date if all refs are rejected
When the `SSHReceivePack()` RPC executes, then we expect it to vote on
all reference updates. While this works alright, we have observed cases
where the RPC doesn't perform any votes at all. This behaviour was
reproduced with one of the testcases introduced in the preceding commit:
if all reference updates got rejected, then the RPC does not fail, but
neither does it cast any votes because there was no reference to be
updated.

One first needs to think about whether rejected updates should be
treated as a successful RPC invocation. From the point of view of
Gitaly, everything works as expected: we did receive reference updates
as well as the packfile from the client, only that some of the updates
aren't accepted because their expected state doesn't match the actual
state. In essence, this is a PreconditionFailed error, and the user will
always get notified about it regardless of whether we fail the RPC or
not. If becomes hairier to judge when only a subset of refs got rejected
and others went through. In the end, it is clear that this is a totally
legit and expected failure which we experience all the time. And if we
would start to treat it as a real failure, then this would bring along
ramifications like scheduling replicaton jobs on partially or fully
rejected pushes, which doesn't make a lot of sense.

The picture would shift a bit if the user did an atomic push, because
then it's clear that he either wants all or nothing. But let's not
consider this scenario for now.

So with that line of reasoning, it feels like the right thing to do to
not raise an RPC error when ref updates get rejected (but naturally, we
report the actual Git errors to the user). With that established, we
must treat it as an error when there are no transactional votes as
otherwise we'd be cautious and schedule replication jobs to secondaries.

The fix is the usual one: after git-receive-pack(1) has finished, we do
one more manual vote. This does serve two purposes: first, we know we
always cast at least one vote regardless of whether something was
upgraded or not. But second, this also makes sure that the sequence of
reference updates on all nodes is the same and is indeed finalized.
Otherwise it may happen that some nodes finish the RPC early in case
they're updating less references, while other nodes hang while waiting
for a quorum on their current vote.

Instead of casting a vote with a fixed value, we could also vote on the
repository's state after git-receive-pack(1). This wouldn't quite
reflect the state we care about though: we only care about what we
actually changed in the RPC, not about the complete repository's state.
So using a fixed value for the vote feels like the correct thing to do.

Changelog: fixed