Do not handover internal socket during upgrades
Gitaly currently hands over its internal socket to the new process
during graceful upgrades. This is problematic as Gitaly now contains
some state that is not transfered to the new process, namely the
backchannel connection identifiers. Gitaly routes votes back to the
source connection of the mutator RPC. It identifies the source connection
by a monotonically increasing identifier, which is local to the process.
If the Praefects connect to the new process in different order than the
old process, their connection identifiers would differ between the old and
the new Gitaly process. This can cause Gitaly to route a vote to wrong
Praefect during upgrades if there are in-flight requests going through
the hooks, which then end up in the new process instead of the old one.
In such a case, the connection identifier may point to a different Praefect.

This commit fixes the problem by not handing over the internal socket to the
new process. The internal sockets are suffixed by the process' ID to avoid
naming collisions. With a unique internal socket for both processes during the
ugprade, the hook calls hit the process that was originally handling the RPC.
This ensure the same process handles the votes as well and that the backchannel
ID refers to the correct Praefect always.

Changelog: fixed