voting: Move Vote and VoteHash types into shared package
The infrastructure for transactional voting is currently duplicated
across Praefect and Gitaly. As a preparatory step to fix this and unify
both, this commit moves Gitaly's `Vote` and `VoteHash` types into a new
agnostic package "internal/transaction/voting".

It would be more idiomatic to call the package "vote" instead of
"voting", or to have the implementation in "internal/transaction". But
the former is creating problems because one cannot name local variables
"vote" anymore, while the latter causes problems because we now have
import conflicts in a lot of places (e.g. "internal/gitaly/transaction"
vs "internal/transaction"). As such, "voting" seems like the best
compromise to avoid both issues.