testing: Assert access rather than permission mask
Given `git-init(1)` might give different permission than the defined
0770 the parent directory of a repo might have, the test might fail
depending on the users system.

This change updates the test to assert read, write, and execution
permissions, rather than checking a mask or the masks String(). That
asserts what we really care about. This assertion is made on the
relevant subdirectories too.