Handover Praefect sockets on upgrade
In bbcf1e23, the `starter` package gained `HandoverOnUpgrade` parameter.
Adding the parameter, I missed that Praefect also is using it, and compiles
fine without adding the parameter to the struct literals. This means Praefect
currently would not handover the sockets during ugprade, which breaks the zero
downtime upgrades. This commit fixes the situation by setting Praefect's sockets
to be handed over during the upgrades.