metadata: Move transaction metadata into shared package
Now that we have a shared `internal/transaction` package, it makes sense
to move over transactional metadata there too given that it's used by
both Gitaly and Praefect. This commit thus moves the package and renames
it to `txinfo`, which gives a lot more context compared to the generic
`metadata` name.