Close internal socket after the external ones
During the graceful shutdown process, Gitaly's internal socket should
close after the external request accepting sockets. If an external
request is accepted into Gitaly and the graceful shutdown is initiated
on all of the sockets, the request will not complete successfully if
it tries to contact the HookService via the internal socket. This is
due to the internal socket already performing a graceful shutdown
as well, causing it to stop accept new RPCs. This can cause requests
to fail during graceful upgrades.

This commit fixes the problem by separating the internal sockets from
the external sockets in the GitalyServerFactory. Doing so, the
GracefulStop method can first close the external sockets and only
after successfully draining them, also close the internal sockets.
This ensures internal socket does not reject any RPCs already in-flight.
When the external sockets complete their graceful shutdown, the
connections should be fully drained and closing the internal sockets is
safe.

Changelog: fixed