Break dependency of the CloneRepoAtStorage on the global config.Config
In order to get rid of the global variable config.Config we change
CloneRepoAtStorage function to accept config.Cfg as an input parameter.
We use locally defined instance of the config.Cfg type to call modified
function, so the usage of it is no longer depend on the config.Config
variable.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699