Remove testhelper.CloneRepo
CloneRepo function relies on the global config.Config variable
and as we are on the road to remove it from the code we drop
CloneRepo function and replaces its usages with an alternative
testcfg.BuildWithRepo function call. The change also includes
renaming: testRepo -> repo, testRepoPath -> repoPath for local
variables.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699