log: Rename file hosting log parser
The git-log parser is currently hosted in a file named "log.go", which
is rather generic and doesn't really tell you a lot about what it hosts
given that it's already located in a package called "log".

Rename the file to "parser.go" to fix this. While at it, drop the
stuttering in `log.NewLogParser()` by renaming the function to
`log.NewParser()`.