log: Move tag and commit parsing into catfile package
The log package contains a parser for both commits and tags. It's quite
weird that this is in the log package: we're parsing objects here, not
the log, and the functions are even called `GetCommitCatfile()` and
similar.

Move this functionality over into the catfile package: the
implementation is deeply tied to it already, so it's only natural to
have it there. Ideally, we'd also implement it on the `Batch` interface.
But right now, this feels like needless additional churn.