Ignore diverged non-matched branches in UpdateRemoteMirror Go port
UpdateRemoteMirror updates a remote push mirror of a repository. If the
KeepDivergentRefs flag is set in the request, UpdateRemoteMirror doesn't
force push over diverged refs but returns them in the response. The Ruby
implementation of the RPC only includes diverged branches that match the
branch selector in the request. However, the Go port of the RPC
currently includes also diverged branches that should be ignored by the
branch selector. This caused some issues when the feature flag was enabled
in production as the RPC reporter mirroring errors for diverged branches
that should not be mirrored in the first place according to the branch
selector.

This commit fixes the issue by only including the diverged branch in the
response if it matches the branch selector.