cache: Rename primary structure to `Cache`
The primary structure of the cache package is currently called
`StreamDB`, which is a bit confusing. Given that no new cache types have
been added to the package over the two years where the disk cache has
existed, let's rename it to `Cache` to make it the primary exported
structure of the package. This also prepares for a world where no other
types will be exported by it because everything else is an internal
implementation detail.