Log a deprecation warning when running reconcile
The manual reconciler is no longer needed as Praefect's automatic
reconciler performs reconciliations automatically, in a safer manner
and with better performance. The manual reconciler works by mirroring
a whole storage which makes it a bad fit with variable replication factor
as the storages are not mirrors of each other anymore. Any jobs the
manual reconciler would schedule but the automatic one wouldn't are dropped
by the checks like downgrade protection. There's really no reason to run
the manual reconciliation anymore so it is deprecated in GitLab 13.12 for
removal in 14.0. Given that, this commit logs a warning message in case
anyone still is using the command.

Changelog: deprecated