Remove on-by-default gitaly_go_user_update_branch feature flag
Remove the the gitaly_go_user_update_branch feature flag that's been
on by default since my 866b08492 (Turn UserUpdateBranch in Go on by
default, 2021-03-23) merged in [1]. This brings us one step[2] closer
to removing this entirely.

As noted in [3] the underlying Ruby code is not being removed in this
commit, we'll have to wait until this change has been out for a while
to avoid the race condition described in that documentation.

1. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/3286
2. https://gitlab.com/gitlab-org/gitaly/-/issues/3472
3. https://gitlab.com/gitlab-org/gitaly/-/blob/master/doc/PROCESS.md#two-phase-ruby-to-go-rollouts

Changelog: changed