blob: Generalize LFS pointer filter to apply to blobs
The LFS pointer filter filters blobs read from a reader and only returns
those which could be potential LFS pointers. This helper can easily be
generalized to filter blobs based on an arbitrary filter -- while not of
use right now, it will eventually be of use when we start to implement a
new RPC to list blobs.

So let's generalize the filter now to be prepared for that future.