server: Use injected Cache as cache invalidator
With the preceding commit, the `Cache` has been converted to be the
primary gatekeeper of leases and is thus the central authority to all
things disk caches. The middleware is still using an ad-hoc created
LeaseKeyer instead of the global cache.

Inject the global cache into the server factory and use it as cache
invalidator.