blob: Use ProtoEqual() to compare LFS pointers
While comparing protobuf messages for equality directly is discouraged,
we still do so in our LFS pointer tests. This has worked alright for
now, but in a subsequent commit we're about to change our implementation
which will cause breakage here.

Refactor the code to use a new function `lfsPointersEqual()` function,
which compares both expected and actual pointers with
`testhelper.ProtoEqual()`.