featureflag: Activate Rebase implementation in Go
The feature flag for the Go implementation for `UserRebaseConfirmable`
is enabled on production for a few days already, and has been without
without any issue.

This change sets the default state to enabled, so the Go implementation
will also be active at all customers, but still give them the option to
disable it.

Related to https://gitlab.com/gitlab-org/gitaly/-/issues/3575

Changelog: performance