Fix lint task on MacOS
When lint task runs on MacOS it reports about redundant
lint error suppressing:
nolintlint  directive `//nolint:unconvert` is unused for linter "unconvert"
It happens because the OS specific structs are used in
the code and there is a different types of the same named
field for MacOS and Unix systems.
We suppress that check by adding nolintlint exclusion.