wiki: Remove GetPageVersionsRPC
There's one client that uses PageVersions, Rails. This RPC ultimately
meant; what versions of a certain path exists for a repository. There
already was an RPC for that, and now Rails uses that one. This means we
can remove this RPC a release after the Rails code has been release.

This change removes most of the dead code, probably not all but the
risk/reward ratio of removing code in the sidecar is high until we
remove the sidecar alltogether.

Upstream changes: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/61459