Respect failover disabled config option with per_repository elector
Praefect has a config option toggling whether the failover logic is
enabled or disabled. When repository-specific primaries are enabled,
the configuration option is not currently respected. This commit fixes
the situation by not running the election loop in case failover is
disabled. Rest of the logic stays the same, and the primaries are still
being read from the database, unlike with `sql` elector with failover
disabled. Praefect just doesn't update the primary records if failovers
are disabled.

Changelog: fixed