Remove GrpcErrorHasMessage test helper
There is no much sense in use of the GrpcErrorHasMessage
helper function. It is better to use equality check for
the received error and expected result as in that case
you check both - the error message and the status code,
so we are more confident about the implementation and
a contract.