cache: Implement Leaser interface for `Cache` structure
It's currently the LeaseKeyer's responsibility to create leases. This is
an awkward interface to use because we cannot reuse information hosted
in the `Cache` structure itself, and arguably it should be the `Cache`'s
responsibility to be the gatekeeper of all leases. The result is that we
have implicitly shared state between leasers and disk caches where both
must be created with the same set of parameters, otherwise we'll run
into hard to debug issues. One location where this shows is the
middleware, which creates another ad-hoc leaser to invalidate caches --
which again, should really be handled by the cache structure itself.

As a preparatory step to disentangle this mess, implement the Leaser
interface for the `Cache` structure by forwarding calls to `StartLease`
to its keyer.