git: Capture GIT_TRACE2_EVENTs
Git allows a caller to receive metrics and events through `GIT_TRACE` and
relatives; this has been around for a while. If needed, one could enable
these before this change with `GIT_TRACE=1 ./gitaly` on boot. While this
works for debugging, it logs a lot of events and thus is not enabled on
production system.

This change introduces a mechanism to enable the second iteration of the
tracing from Git. Again Git exposes a lot of events and metrics, though
now the there's improvements like setting a custom sink, exposing JSON
over unstructured text, and better context among lines emitted.

For now Gitaly creates a new open file descriptor when a Git command is
spawned and copies the input from it to an io.Writer. This is not yet
useful, but does sets up plumping to capture the events. Further, it
would now be trivial to copy all events to stderr during testing if one
wanted to.

Later changes will introduce mechanisms to analyse the stream, and
expose its data structured to be more useful in the GitLab architecture.
Say through Prometheus, tracing, and/or logging.