repository: Remove housekeeping from Cleanup RPC
The `Cleanup()` RPC is run synchronously when a push is received,
creating the risk that a failure here will block users from pushing
their changes. In https://gitlab.com/gitlab-org/gitaly/-/issues/3581 we
found that the stale files and refs checks run as part of housekeeping
may exceed 60 seconds for larger repositories hosted on NFS, causing
pushes to consistently timeout for some self-managed customers.

This commit removes the housekeeping call from `Cleanup()` and moves
the tests performed on housekeeping-specific functionality over to
`GarbageCollect()`, which is now the most frequently executed RPC that
runs housekeeping.

Changelog: fixed