config: update a bin_dir of config.toml.example
A default usage of `make` does compile binaries
in `_build/bin`. To make a `config.toml.example`
be easy to use in a default setting this aligns
`bin_dir` with this change.

If a `make install` is done, a default `DESTDIR=/usr/local/bin`
a `bin_dir` needs to be aligned as well.