Push default branch in the first batch of refspecs in UpdateRemoteMirror
UpdateRemoteMirror pushes to a remote mirror repository. The pushes are
done in batches of 10 refspecs. The Ruby implementation of the RPC ensures
the default branch of the repository is always pushed in the first batch
to ensure anything that relies on the default branch works. The Go port
of the RPC is currently not doing so. This commit fixes the situation by
ensuring the default branch is always part of the first push.