Remove unneeded shouldSucceed fields in tests
The coordinator's transaction tests use the shouldSucceed field to
indicate whether a node should successfully vote or not. It's only
used if the node participates in the transaction as indicated by the
shouldParticipate field. This commit removes some extraneous
shouldSucceeds in preparation for refactoring the tests cases to support
multiple subtransactions.