Rename file with package external tests
handler_test.go contains tests in proxy_test package. As this is an
external package, it does not allow us to test any unexported functions
from the proxy package. This commit renames handler_test.go to handler_ext_test.go
so we can use handler_test.go for tests internal to the package.