Mark nodes as failed if the proxy fails opening streams
The transaction finalizer is gathering information about the nodes which
failed the transaction via the ErrHandler passed in with each node. If proxying
an RPC fails already on opening streams to the Gitalys, the errors are not passed
to the transaction finalizer via the ErrHandler. This can cause failed requests to
look like they were successful. This commit amends the situation by failing all the
nodes with any error that occurs during opening the streams.