server: Inject logger into server factory
While it's possible to specify a specific logger via `server.New()`, the
server factory will always use the default logger. As a result, we
cannot use the server factory in case we want to exercise logging code
in any of our tests.

Fix this shortcoming by injecting the logger into the server factory
such that it can be set up by the caller.